pyproject.toml
[tool.poetry]
name = "text-to-image-encoder"
version = "0.1.0"
description = ""
authors = ["LexXeno <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.1.1"
random2 = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
